requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
xlrd>=2.0.1
beautifulsoup4>=4.9.3
requests>=2.28.1
Pillow>=9.0.0